+++ src/backend/typeDefs/mutations/teams/create.js
@@ -2,8 +2,10 @@ const { gql } = require('apollo-server-express')
 
 const createTeamTypeDefs = gql`
   input CreateTeamInput {
+    _id: String # not required but same form is used by update team so this comes in as null and needs typeDef
     description: String!
     clientId: String!
+    defaultLandingPath: String!
   }
 
   type CreateTeamPayload {
@@ -14,6 +16,7 @@ const createTeamTypeDefs = gql`
     sitemap: JSON
     client: Client
     resources: JSON
+    defaultLandingPath: String
   }
 `